Separate office common info into Components
- 사무실 공통 정보를 컴포넌트로 분리하다
- `주소`, `이메일`, `전화번호`
