+++ components/utilities/officeEmail.js
@@ -1,0 +1,5 @@
+function SaemmulterEmail() {
+  return <a href='mailto:cgwaterlb@gmail.com'>cgwaterlb@gmail.com</a>
+}
+
+export default SaemmulterEmail;
