+++ components/utilities/officeTel.js
@@ -1,0 +1,5 @@
+function SaemmulterTel() {
+  return <a href='tel:+82-10-7413-1102'>+82-10-7413-1102</a>
+}
+
+export default SaemmulterTel;
